Add initialization functions; update signatures
ncurses organizes its functions and names into categories based on
shared purpose. This change registers all the functions in the
"initialization" category, as described in the curs_initscr man page.

I've also changed the types of `initscr` and `endwin` functions to
directly reflect the signatures of the ncurses functions (initscr
returns a pointer to a WINDOW, endwin returns an int indicating the
success or  failure of the endwin call).