Register the ERR and OK macros
The ncurses library makes extensive use of two macros to communicate
success and failure for side-effecting routines that may fail, these
macros are OK and ERR respectively.

We register them here so that we can make direct comparisons to the
values of these macros in calls to functions that return them. For
example, `addstr` returns ERR on failure. In our carp code, we can now
compare the result of an `addstr` call to `ERR` directly and determine
how our program should proceed based on the comparison.