Register addch functions
This change registers the ncurses function family for adding individual
characters to the screen. Note that this does not include support for
wide characters, so, characters larger than a byte will likely cause
errors. We'll implement wide character support later.

Note that we also need to register the `chtype` type, defined in
curses.h, to support these functions.