Add attr function family and Attribute type
This change adds the family of attr functions, described in curs_attr.
There are two important points to note about these bindings:

- We need to introduce a new type, `Attribute` to mirror the curses
attr_t type. We can't use attr_t directly because on the Carp side, we
need to wrap attributes (attr_t values) in the `Ptr` constructor to
denote pointers. Lowercase values passed to constructors such as Ptr are
interpreted as type variables, not the concrete type we want here, so we
need to use an alias.

- Several of these functions take void pointers as a final argument;
most of them are unused, excepting functions that can change colors
(according to curs_attr) which instead expect an Int pointer; so,
instead of using a `(Ptr ())` type, we set the binding signatures to
`(Ptr Int)` and pass Carp's NULL wherever an Int isn't supported in
ncurses.