Conditionally define cchar_t
The cchar_t struct type is only defined when NCURSES_WIDECHAR is
defined, which, in turn, is only defined if _XOPEN_SOURCE_EXTENDED is
defined--the general effect of this is to include unicode support in
ncurses (one also has to use a separate set of wide character routines).

We'll add better support for this in the future.