Add Int->Chtype converter; encapsulate conversions
I've updated the example according to new function signatures.

In order to match the bindings, I had to introduce a conversion function
from Ints to Chtype values. Since we have an equivalent conversion from
Int to Short, I've enclosed each conversion function in a submodule of
the NCurses module (so that we don't collide with other possible types
conversion functions outside of the module).